Server/requirements.txt
paho-mqtt
matplotlib
numpy
colorama